(PE-339) Move license check to cert sign/revoke
Prior to this commit, the license check did not execute as part of
certificate sign and revoke, the actions most likely to affect license
status. This commit causes certificate signing and revoking to initiate
a license check (in a new non-blocking thread for the puppet master).

This commit is an application of a patch by @jeffweiss, retrieved from
https://github.com/puppetlabs/pe-rpm-puppet/pull/8

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>