Revert "(#15187) Ensure that fatal errors are logged to console"
This reverts commit 587d942ce975b0ba5f566c71a3f5b78650deabb8.  This commit
causes problems because it causes the agent service to hang onto stdout after
it has daemonized, so anything wrapped on top of that service will also hang
waiting for the agent to let go of stdout. The PE installer tees the service
and this commit causes the installer to hang on service start.