Combine AIX and rhel spec files
This commit moves all of the logic from the AIX pe-puppet spec file into
the EL/SLES spec file. This is to solve two problems:

1) The current setup maintains a separate, non-templated spec file for
pe-puppet in the pe-aix-rpm-puppet repository. Every time we update
pe-puppet, we have to update this file manually with new versions. By
combining these, the spec file will now get written by the packaging
repo and AIX can just use it like the other platforms.

2) When pe-puppet moved to the packaging repo for EL/SLES/Debian, the
versioning became dynamic based on what is in the git repo, but for
Solaris, AIX, and Windows, this is still not the case. This means to get
the version we put into the AIX specfile, we have to open the tarball
and get the version from the redhat spec that we're going to use. By
combining these, this no longer an issue - we can just use the redhat
spec on AIX.

It also removes an obsolete %package server requirement on activerecord.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>