Disable SSL verification for http report processor
This is used to connect to the console, which has the certname
pe-internal-dashboard rather than its hostname. For historical reasons,
that cert doesn't have alt names, so this will fail now that we use the
standard http report processor which uses SSL verification where the
old https report processor did not.

In future, either we should fix that cert or just stop sending reports
to the console. For now, we regress to the previous behavior of not
verifying the connection.