(PE-566) Fix upgrade of hiera packages
Prior to this commit, the pe-puppet package only obsoleted version of
pe-rubygem-hiera-puppet < 1.0.0, which is great, except that the latest
that we've shipped is 1.0.0-2, causing the new pe-puppet package to no
obsolete these old hiera packages. This commit updates the version that
pe-puppet provides to be 1.0.0-3 (one higher than previously shipped) so
that rpm can properly upgrade the packages.