PE-615: Don't create C:/dev/null on Windows
The puppet apply application uses the "store" report processor by default.
Its process method calls mkdir_p on Puppet[:reportdir]. When executing in a
spec test, Puppet[:reportdir] is '/dev/null/reports'. On Unix, the call to
mkdir_p raises an error, which is ignored (since the test is exercising
apply functionality, not the store report processor). But on Windows, the
call to mkdir_p succeeds.

The existence of the c:/dev/null directory causes bundler to fail, and is
why the windows spec jobs do `rmdir /S /Q C:\dev\null`.

This commit updates the tests to use the "none" report processor, which
skips report processing altogether, since it's not germane to the test.