(maint) Remove rspec requires from the Rakefile
In the Rakefile, we no longer use Rspec methods or classes, but instead shell
out to rspec, so requiring rspec serves no purpose in the rakefile. The require
was also incorrectly placed within an unrelated begin/rescue block, which
caused deprecated rake libraries to be loaded when rspec was not present. This
commit removes it from the Rakefile.