(PE-1644) Prevent pe-puppet from starting agent on upgrade
Prior to this commit, during upgrades on debian systems, debhelper would
kindly come in and help us with our upgrades. It would start pe-puppet,
even though we don't want the arent to be running yet. This commit takes
debhelper out of the picture in the postinst. This makes it so debhelper
doesn't add a lovely block of code at the bottom of the
pe-puppet.postinst that starts the service when that file is loaded.