(RE-522) Puppet should create its own symlinks
Prior to this commit, the installer would create symlinks for binaries
that we want available in /usr/local/bin. Since we're creating the
frictionless agent install, which doesn't take adventage of the
functionality of the installer, these symlinks are not created. As a
fix, we need the packages themselves to create their own symlinks from
/opt/puppet/bin to /usr/local/bin