(bug) Fix reports dir listed multiple times
Both the reports dir and it's parent dir are listed
in the RPM spec's %files section. As listing a dir
in %files (without the %dir directive) also includes
all of the files and directories under that directory
this results in an error due to the reports dir being
accounted for twice.

This commit removes the redundant and unnecessary
declaration of the reports directory.