(QA-786) Remove field based search tests
QA-786-pmt-remove-field-based-search-tests

Remove tests based on determining if the PMT/Forge is honoring
searches by specific fields. (description, keyword, module name).
The docs and man pages provide no expectation to the user that
module searching can be performed by querying specified fields.

Change the strings that are matched during the search UX
interaction to be more general and reflect the current strings
in the responses as of SHA f4f5446 of pe-puppet.