(PF-382) Bypass PE version checks using --force
PE version requirements are implemented with a metadata field
indicating the versions of PE that are supported.  Previously, if any
release of a module indicated its support of the current PE version,
only those releases which supported the current PE version were
considered for dependency resolution.  This was true both with and
without the use of the `--force` flag.

Since the intent of the `--force` flag is to bypass all such checks,
this change fixes that bug by ensuring that such filtering only takes
place when the flag has not been passed.