(PF-384) Installed modules ignore PE version requirements.
It was possible to force-install a module with PE version requirements
that were not unmet (which is reasonable), but installing modules that
depended on that module would either fail to resolve (bad) or force an
upgrade (which may not have been necessary).  This change bypasses the
PE version check for presently installed modules, allowing them to be
used as dependencies without necessarily meeting the global criteria.