(PE-2899) Fix `puppet module list --tree` for modules without metadata.
Previously, trying to build a tree for modules without metadata caused an exception to be raised when trying to look up dependency information for that module.  This was owing to the fact that the dependency hash had a default proc that made certain assumptions about the format of the key, which were valid when the hash was being generated, but not outside of that context.

The fix demonstrated here was simply removing that default proc.