(PF-446) Add a PMT option to ignore incompatibile releases.
Previously, if you wanted to install a module release that expressed compatibility with a PE version you didn't have, the only option available to you was to install that version with `--force`.  This had the unfortunate side-effect of ignoring several other inbuilt protections, possibly leaving your system in a less functional state than anticipated.

This patch adds the `--ignore-incompatibility` option, which (as the name implies) will permit modules to be installed and upgraded despite describing themselves as incompatible with your version of PE.