(maint) Add pe-hiera as a build-depends
In commit 8b482197744d0f45e6da0c6e890abb6844852d8f, a check was added to
install.rb to verify that hiera is present when installing. Because the
packaging uses install.rb to install puppet, this means that it must be
available at build time and must be added as a
BuildRequires/Build-Depends. This commit does that.