(PDB-743) Workaround PuppetDB bug PDB-707 for Puppet PE tests
This patch adds a simple curl request to trigger a 57P01 exception early to
avoid the 'first time seen' database disconnection error bug described in
PDB-707.

For some reason there is a race condition in PE that causes PostgreSQL to
restart underneath PuppetDB triggering PDB-707 in the PE Puppet tests. There
is lots of discussion in that ticket on how to fix that issue, but in the
meantime we need a work around for Puppet tests.

This patch provides such a workaround, and appropriate mention of the real
issue in the test file so that we can always lay blame for this.

I've always somewhat improved the handling of the JSON failure so we get
a dump of the originating text error message, since the JSON exception would
ordinarily swallow this. This may help us in the future debug other issues
so I believe its worth the minor fix, even if it didn't help us this time.

Signed-off-by: Ken Barber <ken@bob.sh>