(PE-5547) Flip stringify_facts to 'false'
Previously, stringify_facts inherited the upstream value of 'true',
to convert all facts into strings. This had the effect of negating
the use of structured facts in Facter, which is undesirable in PE3.4.

This commit changes the default to 'false', allowing structured facts
to pass through to the server in their natural state.