(PE-5548) Set 'trusted_node_data' default to 'true'
Prior to this commit, 'trusted_node_data' was 'false', to match
the upstream Puppet FOSS default. This does not enable the
top-scope '$facts' and '$trusted' hashes, because users who
have those variables defined would see errors upon upgrading,
so we're preserving these for FOSS Puppet 4.

However, not enabling them in PE3.4 breaks the end-to-end
availability of structured facts and is a relatively low-risk
change, so pe-puppet should flip the default before upstream.