(PE-5331) Remove external/ copy of semantic gem
In the original use of this gem in pe-puppet for PE 3.2, it was added to
external/.  When the code was ported to FOSS puppet, it was added to
vendor/. That change was then merged into pe-puppet, resulting in two
copies of the gem being around. Since the external/ version is no longer
used, it ought to be removed.