PE-5724 Improve the spec tests around PE version
Before this commit, the spec tests check to see if the PE version is a
SemVer. It is never expected to be, so this is a broken spec test. After
this commit, it does not check for SemVer when running PE. (Note this is
a pe-puppet commit, because in order to check if it is enterprise, one
must be running pe-puppet.)